Use Capabilities Request command to detect displays
Until now the Identification Request command was used to detect
devices. The command is not strictly mandatory for displays to
implement but should at least result in a valid error response.

Some devices seen to not even send back an error/null response
but instead only respond with null bytes causing the detection
to abort.

Now the first chunk of the capability string is requested for
the first detection step. As the capabilites request command is
effectively mandatory, this should improve compatibility with
badly programmed displays.

May fix issues #11 and #20.